restructured test files and removed stale test
